move pkg/stringsearch.FieldNameFinder to runtime/expr
pkg/stringsearch isn't a great home for FieldNameFinder, which isn't
concerned with general string seach.  Move it to runtime/expr, where
it's used.
